feat(entity): 회원 관련된 엔티티들 추가
