+++ src/main/java/com/nhnacademy/back/account/customer/domain/entity/Customer.java
@@ -1,0 +1,29 @@
+package com.nhnacademy.back.account.customer.domain.entity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Customer {
+
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private long customerId;
+
+	@Column(nullable = false, length = 100)
+	private String customerEmail;
+
+	@Column(nullable = false, length = 20)
+	private String customerPassword;
+
+	@Column(nullable = false, length = 20)
+	private String customerName;
+}
