+++ src/main/java/com/nhnacademy/back/account/memberrole/domain/entity/MemberRole.java
@@ -1,0 +1,24 @@
+package com.nhnacademy.back.account.memberrole.domain.entity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.Getter;
+
+@Getter
+@Entity
+public class MemberRole {
+
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private long memberRoleId;
+
+	@Column(nullable = false)
+	@Enumerated(EnumType.STRING)
+	private MemberRoleName memberRoleName;
+
+}
