+++ src/main/java/com/nhnacademy/back/account/memberrole/domain/entity/MemberRoleName.java
@@ -1,0 +1,5 @@
+package com.nhnacademy.back.account.memberrole.domain.entity;
+
+public enum MemberRoleName {
+	ADMIN, MEMBER
+}
