+++ src/main/java/com/nhnacademy/back/account/memberstate/domain/entity/MemberStateName.java
@@ -1,0 +1,4 @@
+package com.nhnacademy.back.account.memberstate.domain.entity;
+
+public enum MemberStateName {
+}
