+++ src/main/java/com/nhnacademy/back/account/rank/domain/entity/RankName.java
@@ -1,0 +1,5 @@
+package com.nhnacademy.back.account.rank.domain.entity;
+
+public enum RankName {
+	NORMAL, ROYAL, GOLD, PLATINUM
+}
