+++ src/main/java/com/nhnacademy/back/account/socialauth/domain/entity/SocialAuthName.java
@@ -1,0 +1,5 @@
+package com.nhnacademy.back.account.socialauth.domain.entity;
+
+public enum SocialAuthName {
+	PAYCO, WEB
+}
