--- src/main/java/com/nhnacademy/back/address/domain/entity/Address.java
@@ -1,46 +1,0 @@
-package com.nhnacademy.back.address.domain.entity;
-
-import java.time.LocalDateTime;
-
-import com.nhnacademy.back.member.domain.entity.Member;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.ManyToOne;
-import lombok.Getter;
-
-@Getter
-@Entity
-public class Address {
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private long addressId;
-
-	@Column(nullable = false)
-	private String addressName;
-
-	@Column(nullable = false, length = 5)
-	private String addressCode;
-
-	private String addressInfo;
-
-	@Column(nullable = false)
-	private String addressExtra;
-
-	@Column(length = 20)
-	private String addressAlias;
-
-	@Column(nullable = false)
-	private boolean addressDefault = false;
-
-	@Column(nullable = false)
-	private LocalDateTime addressCreatedAt;
-
-	@ManyToOne(optional = false)
-	private Member member;
-
-}
