--- src/main/java/com/nhnacademy/back/memberrole/domain/entity/MemberRoleName.java
@@ -1,5 +1,0 @@
-package com.nhnacademy.back.memberrole.domain.entity;
-
-public enum MemberRoleName {
-	ADMIN, MEMBER
-}
