--- src/main/java/com/nhnacademy/back/memberstatus/domain/entity/MemberState.java
@@ -1,21 +1,0 @@
-package com.nhnacademy.back.memberstatus.domain.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.Getter;
-
-@Getter
-@Entity
-public class MemberState {
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private long memberStateId;
-
-	@Column(nullable = false)
-	private MemberStateName memberStateName;
-
-}
