--- src/main/java/com/nhnacademy/back/memberstatus/domain/entity/MemberStateName.java
@@ -1,4 +1,0 @@
-package com.nhnacademy.back.memberstatus.domain.entity;
-
-public enum MemberStateName {
-}
