--- src/main/java/com/nhnacademy/back/rank/domain/entity/RankName.java
@@ -1,5 +1,0 @@
-package com.nhnacademy.back.rank.domain.entity;
-
-public enum RankName {
-	NORMAL, ROYAL, GOLD, PLATINUM
-}
