--- src/main/java/com/nhnacademy/back/socialauth/domain/entity/SocialAuthName.java
@@ -1,5 +1,0 @@
-package com.nhnacademy.back.socialauth.domain.entity;
-
-public enum SocialAuthName {
-	PAYCO, WEB
-}
